dynamic filter bug fixed on restaurant

+++ backend/src/main/java/com/cyro/craveKart/models/Restaurant.java
@@ -63,7 +63,7 @@ public class Restaurant {
 
   private String openingHours;
 
-  private int ratingCount;
+  private Double ratingCount;
 
   @ElementCollection
   @CollectionTable(
